Azure_AI/requirements.txt
# Azure AI Foundry & Azure Services
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
azure-search-documents>=11.4.0
azure-monitor-opentelemetry>=1.2.0
azure-communication-email>=1.0.0
azure-ai-documentintelligence>=1.0.0

# LangChain & LangSmith
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langsmith>=0.1.0

# Databricks
databricks-sdk>=0.18.0

# Machine Learning
tensorflow>=2.15.0
numpy>=1.24.0
scikit-learn>=1.4.0

# OpenTelemetry
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation>=0.43b0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
